requirements.txt
# Django Core API Gateway Requirements

# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.7
django-redis==5.4.0

# API Documentation
drf-spectacular==0.26.5

# CORS and Security
django-cors-headers==4.3.1

# Filtering and Search
django-filter==23.3

# Health Checks
django-health-check==3.17.0

# HTTP Requests
requests==2.31.0

# System Monitoring
psutil==5.9.6

# Development and Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Code Quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Environment Variables
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0 